Provider Registration method updated
