+++ app/Http/Controllers/ProviderAuth/TokenController.php
@@ -57,11 +57,11 @@ class TokenController extends Controller
                 ]);
                 $data = $Provider->toArray();
                 $prov = Provider::find($data['id']);
-                $prov->status = 'approved';
-                $prov->save();
+                //$prov->status = 'approved';
+                //$prov->save();
               //  return response()->json(['data' => $prov]);
 
-                $this->update_provider_services($prov->id);
+              //  $this->update_provider_services($prov->id);
                 if($request->social_unique_id != ""){
                   $prov->social_unique_id = $request->social_unique_id;
                   $prov->confirmation = 1;
@@ -107,24 +107,24 @@ class TokenController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-     public function update_provider_services($provider_id){
-       try{
-       $provider = Provider::find($provider_id);
-       $provider_service = new ProviderService;
-       $provider_service->provider_id = $provider_id;
-       $provider_service->service_type_id = 1;
-       $provider_service->status = 'active';
-       $provider_service->save();
-       return true;
-     }
-     catch (QueryException $e) {
-         if ($request->ajax() || $request->wantsJson()) {
-             return response()->json(['error' => 'Something went wrong, Please try again later!'], 500);
-         }
-         return abort(500);
-     }
+    //  public function update_provider_services($provider_id){
+    //    try{
+    //    $provider = Provider::find($provider_id);
+    //    $provider_service = new ProviderService;
+    //    $provider_service->provider_id = $provider_id;
+    //    $provider_service->service_type_id = 1;
+    //    $provider_service->status = 'active';
+    //    $provider_service->save();
+    //    return true;
+    //  }
+    //  catch (QueryException $e) {
+    //      if ($request->ajax() || $request->wantsJson()) {
+    //          return response()->json(['error' => 'Something went wrong, Please try again later!'], 500);
+    //      }
+    //      return abort(500);
+    //  }
 
-     }
+    // }
     public function authenticate(Request $request)
     {
         $this->validate($request, [
